+++ CHANGELOG.rst
@@ -2,7 +2,10 @@ Changelog
 =========
 
 
+- Replace internal implementation of Unicode date pattern formatting with Babel's ``format_datetime``. **breaking change**
+- Remove support for formatting to timestamp using ``X`` and ``XX``. **breaking change**
 - Add ``zulu.create`` as factory function to create a ``zulu.Zulu`` instance.
+- Add locale support to ``Zulu.format`` when using Unicode date pattern format tokens.
 
 
 v0.5.0 (2016-08-13)